refactor: place Next server reference info adapter
