+++ docs/nextjs-fidelity-architecture-tracker.md
@@ -285,8 +285,8 @@ higher owner is blocked.
 | `plugin/builtin-global-error.test.ts`         | `src/client/components/builtin/global-error.test.ts`                                      | Same                                              | Done in this worktree. Test follows implementation.                                                                                                                                                                                                                  |
 | `plugin/entry-base-client-references.ts`      | `src/server/app-render/entry-base.ts`                                                     | Next app-render entry-base client refs            | Done in this worktree. Vite bridge preserves Next entry-base client reference boundaries. P1 deletion target: let `@vitejs/plugin-rsc` preserve CJS `"use client"` dependency boundaries during RSC dep optimization.                                                |
 | `plugin/entry-base-client-references.test.ts` | `src/server/app-render/entry-base.test.ts`                                                | Same                                              | Done in this worktree. Test follows implementation.                                                                                                                                                                                                                  |
-| `plugin/server-reference-info.ts`             | `src/shared/lib/server-reference-info.ts`                                                 | Next server reference info helper                 | Move mirror behavior there.                                                                                                                                                                                                                                          |
-| `plugin/server-reference-info.test.ts`        | `src/shared/lib/server-reference-info.test.ts`                                            | Same                                              | Test follows implementation.                                                                                                                                                                                                                                         |
+| `plugin/server-reference-info.ts`             | `src/shared/lib/server-reference-info.ts`                                                 | Next server reference info helper                 | Done in this worktree. Virtual helper delegates real hex IDs to Next and preserves Vite RSC module IDs. P1 deletion target: remove when Vite RSC action IDs no longer flow through Next's reducer arg-pruning path.                                                  |
+| `plugin/server-reference-info.test.ts`        | `src/shared/lib/server-reference-info.test.ts`                                            | Same                                              | Done in this worktree. Test follows implementation.                                                                                                                                                                                                                  |
 | `plugin/runtime-rewrites.ts`                  | `src/client/app-index.tsx` or `src/server/web/adapter.ts` depending on content            | Runtime/browser bootstrap rewrites                | Inspect content before move; must name exact source.                                                                                                                                                                                                                 |
 | `plugin/runtime-rewrites.test.ts`             | Follow implementation                                                                     | Same                                              | Test follows implementation.                                                                                                                                                                                                                                         |
 | `plugin/tester-html.ts`                       | Keep top-level plugin plumbing unless source-linked                                       | Vitest tester HTML generation                     | Not necessarily a Next mirror. If it parses app-index bootstrap, split that part to `src/client/app-index.tsx`.                                                                                                                                                      |
